requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
cairocffi==1.6.1
CairoSVG==2.7.1
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.0.2
django-appconf==1.0.6
django-bootstrap-breadcrumbs==0.9.2
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-filter==23.5
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mobile==0.7.0
django-ordered-model==3.7.4
django-paypal==2.0
django-tables2==2.7.0
django-view-breadcrumbs==2.5.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.6
kombu==5.3.5
openpyxl==3.1.2
pilkit==3.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==48
webencodings==0.5.1
wheel==0.42.0
